chess_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests
python-chess==1.999
pytest
pytest-cov
pytest-mock
supabase==2.1.0
python-dotenv
httpx==0.24.1
black
isort
flake8
mypy
types-requests
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
